+++ src/uci.rs
@@ -150,7 +150,7 @@ fn bench(pos: &mut Position, pos_data: &Arc<RwLock<PosData>>, args: &str) {
 
     let duration = now.elapsed();
     let elapsed =
-        (duration.as_secs() as u64) * 1000 + (duration.subsec_nanos() as u64) / 10000000 + 1;
+        duration.as_secs() * 1000 + (duration.subsec_nanos() as u64) / 10000000 + 1;
 
     eprintln!(
         "\n===========================\
